% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{adj_rand_index}
\alias{adj_rand_index}
\title{Compute Adjusted Rand Index for clustering accuracy.}
\usage{
adj_rand_index(x, y)
}
\arguments{
\item{x}{First clustering vector.}

\item{y}{Second clustering vector.}
}
\description{
Compute Adjusted Rand Index for clustering accuracy.
}
