% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{best_label_accuracy}
\alias{best_label_accuracy}
\title{Compute best label accuracy over all permutations.}
\usage{
best_label_accuracy(true, est, K)
}
\arguments{
\item{true}{True cluster labels.}

\item{est}{Estimated cluster labels.}

\item{K}{Number of clusters.}
}
\description{
Compute best label accuracy over all permutations.
}
