% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{chol_spd}
\alias{chol_spd}
\title{Helper function to compute a Cholesky factor for a numerically symmetric
positive definite matrix.}
\usage{
chol_spd(Sigma)
}
\arguments{
\item{Sigma}{Covariance matrix.}
}
\description{
Helper function to compute a Cholesky factor for a numerically symmetric
positive definite matrix.
}
