% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{cluster_stats}
\alias{cluster_stats}
\title{Helper function to compute cluster counts and cluster-specific sums.}
\usage{
cluster_stats(Z, Y, K)

cluster_stats(Z, Y, K)
}
\arguments{
\item{Z}{Vector of cluster allocations.}

\item{Y}{Matrix of observations.}

\item{K}{Number of clusters.}
}
\description{
Helper function to compute cluster counts and cluster-specific sums.

Helper function to compute cluster counts and cluster-specific sums.
}