% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{ffbs_one_cluster}
\alias{ffbs_one_cluster}
\title{Forward-filtering backward-sampling for one cluster under the
Polya-Gamma Gaussian representation.}
\usage{
ffbs_one_cluster(N, S, omega_k, Fmat, m0, C0, G, W, r_tune)
}
\arguments{
\item{N}{Cluster size.}

\item{S}{Cluster-specific sufficient statistics.}

\item{omega_k}{Polya-Gamma auxiliary variables for cluster k.}

\item{Fmat}{Design matrix (TT x p).}

\item{m0}{Prior mean vector.}

\item{C0}{Prior covariance matrix.}

\item{G}{State transition matrix.}

\item{W}{State evolution covariance.}

\item{r_tune}{Tuning parameter for Polya-Gamma augmentation.}
}
\description{
Forward-filtering backward-sampling for one cluster under the
Polya-Gamma Gaussian representation.
}
