% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{get_eta}
\alias{get_eta}
\title{Helper function to compute eta from theta and the design matrix.}
\usage{
get_eta(theta, Fmat)
}
\arguments{
\item{theta}{Array of parameters.}

\item{Fmat}{Design matrix.}
}
\description{
Helper function to compute eta from theta and the design matrix.
}
