% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_lambda}
\alias{get_lambda}
\title{Helper function to compute lambda from theta and the design matrix.}
\usage{
get_lambda(theta, Fmat)

get_lambda(theta, Fmat)
}
\arguments{
\item{theta}{Array of parameters.}

\item{Fmat}{Design matrix.}
}
\description{
Helper function to compute lambda from theta and the design matrix.

Helper function to compute lambda from theta and the design matrix.
}