% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{logsumexp}
\alias{logsumexp}
\title{Stable computation of the log-sum-exp function.}
\usage{
logsumexp(x)

logsumexp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Stable computation of the log-sum-exp function.

Stable computation of the log-sum-exp function.
}