% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{mode_int}
\alias{mode_int}
\title{Compute the mode of an integer vector.}
\usage{
mode_int(x, K)
}
\arguments{
\item{x}{Integer vector.}

\item{K}{Maximum possible value (number of bins).}
}
\description{
Compute the mode of an integer vector.
}
