% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{plot.bdynets_mcmc}
\alias{plot.bdynets_mcmc}
\title{Plot MCMC diagnostics for bdynets_mcmc object.}
\usage{
\method{plot}{bdynets_mcmc}(x, burn = NULL, type = "all", ...)
}
\arguments{
\item{x}{Object of class "bdynets_mcmc".}

\item{burn}{Number of burn-in iterations to discard.}

\item{type}{Type of plot: "trace", "density", "ari", or "all".}

\item{...}{Additional arguments passed to plot.}
}
\description{
Plot MCMC diagnostics for bdynets_mcmc object.
}
