% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{rdirichlet1}
\alias{rdirichlet1}
\title{Helper function to sample from a Dirichlet distribution.}
\usage{
rdirichlet1(alpha)

rdirichlet1(alpha)
}
\arguments{
\item{alpha}{Numeric vector of Dirichlet concentration parameters.}
}
\description{
Helper function to sample from a Dirichlet distribution.

Helper function to sample from a Dirichlet distribution.
}