% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{relabel_by_mean_intensity}
\alias{relabel_by_mean_intensity}
\title{Label-switching correction via ordering by average fitted intensity.}
\usage{
relabel_by_mean_intensity(theta, pi_vec, Z, Fmat)
}
\arguments{
\item{theta}{State parameters (K x TT x p).}

\item{pi_vec}{Cluster probability vector.}

\item{Z}{Cluster allocation vector.}

\item{Fmat}{Design matrix (TT x p).}
}
\description{
Simple identifiability constraint: order clusters by average
fitted intensity. This is suitable when true clusters are generated
with increasing average intensity.
}
