% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{rmvn}
\alias{rmvn}
\title{Helper function to sample from a multivariate normal distribution.}
\usage{
rmvn(mu, Sigma)
}
\arguments{
\item{mu}{Mean vector.}

\item{Sigma}{Covariance matrix.}
}
\description{
Helper function to sample from a multivariate normal distribution.
}
