% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{rpg}
\alias{rpg}
\title{Approximate Polya-Gamma sampler.}
\usage{
rpg(n, h, z, trunc = 80)
}
\arguments{
\item{n}{Number of samples.}

\item{h}{Shape parameter.}

\item{z}{Location parameter.}

\item{trunc}{Truncation level for approximation. Defaults to 80.}
}
\description{
If BayesLogit is installed, this uses BayesLogit::rpg().
Otherwise, it uses a truncated infinite-sum approximation.
}
