% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{sample_prior_theta}
\alias{sample_prior_theta}
\title{Sample a prior state trajectory from a Dynamic Linear Model (DLM).}
\usage{
sample_prior_theta(TT, m0, C0, G, W)
}
\arguments{
\item{TT}{Number of time steps.}

\item{m0}{Prior mean vector for the initial state.}

\item{C0}{Prior covariance matrix for the initial state.}

\item{G}{State transition matrix.}

\item{W}{State evolution covariance matrix.}
}
\value{
A matrix of size TT x p containing the sampled state trajectory.
}
\description{
Simulates a state sequence \eqn{\theta_{1:T}} from a Gaussian DLM defined by:
\deqn{\theta_0 \sim N(m_0, C_0)}
\deqn{\theta_t = G \theta_{t-1} + w_t, \quad w_t \sim N(0, W)}
}
