% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{sample_theta_all}
\alias{sample_theta_all}
\title{Sample all state trajectories using FFBS.}
\usage{
sample_theta_all(Y, Z, theta, Fmat, m0, C0, G, W, r_tune, K)
}
\arguments{
\item{Y}{Response matrix (n x TT).}

\item{Z}{Cluster allocation vector.}

\item{theta}{Current state parameters (K x TT x p).}

\item{Fmat}{Design matrix (TT x p).}

\item{m0}{Prior mean vector.}

\item{C0}{Prior covariance matrix.}

\item{G}{State transition matrix.}

\item{W}{State evolution covariance.}

\item{r_tune}{Tuning parameter for Polya-Gamma augmentation.}

\item{K}{Number of clusters.}
}
\description{
Sample all state trajectories using FFBS.
}
