% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{summary.bdynets_mcmc}
\alias{summary.bdynets_mcmc}
\title{Summarize MCMC results from gibbs_sampler.}
\usage{
\method{summary}{bdynets_mcmc}(object, burn = NULL, ...)
}
\arguments{
\item{object}{Object of class "bdynets_mcmc".}

\item{burn}{Number of burn-in iterations to discard.}

\item{...}{Additional arguments (not used).}
}
\description{
Summarize MCMC results from gibbs_sampler.
}
