% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{update_Z_given_theta}
\alias{update_Z_given_theta}
\title{Sample cluster allocations Z_i conditional on theta.}
\usage{
update_Z_given_theta(Y, theta, Fmat, pi_vec)
}
\arguments{
\item{Y}{Response matrix (n x TT).}

\item{theta}{State parameters (K x TT x p).}

\item{Fmat}{Design matrix (TT x p).}

\item{pi_vec}{Cluster probability vector.}
}
\description{
Sample cluster allocations Z_i conditional on theta.
}
