% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{bdynets-package}
\alias{bdynets}
\alias{bdynets-package}
\title{bdynets}
\description{
This package implements the \code{bdynets} method.
}
\author{
\strong{Maintainer}: Sheng Jiang \email{jiangsheng@cuhk.edu.cn}

}
\keyword{internal}