% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{logsumexp}
\alias{logsumexp}
\title{Stable computation of the log-sum-exp function.}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Stable computation of the log-sum-exp function.
}
