% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{rmvn}
\alias{rmvn}
\title{Helper function to sample from a multivariate normal distribution.}
\usage{
rmvn(mu, Sigma)

rmvn(mu, Sigma)
}
\arguments{
\item{mu}{Mean vector.}

\item{Sigma}{Covariance matrix.}
}
\description{
Helper function to sample from a multivariate normal distribution.

Helper function to sample from a multivariate normal distribution.
}