% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-helpers.R
\name{sample_omega_all}
\alias{sample_omega_all}
\title{Sample omega_{k,t} once per MCMC sweep.}
\usage{
sample_omega_all(Y, Z, theta, Fmat, r_tune, K)
}
\arguments{
\item{Y}{Response matrix (n x TT).}

\item{Z}{Cluster allocation vector.}

\item{theta}{State parameters (K x TT x p).}

\item{Fmat}{Design matrix (TT x p).}

\item{r_tune}{Tuning parameter for Polya-Gamma augmentation.}

\item{K}{Number of clusters.}
}
\description{
Sample omega_{k,t} once per MCMC sweep.
}
